[CHERRY-PICK] bpf: use user pointer types in uAPI structs
Cherry pick "b11a04266088 bpf: use user pointer types in uAPI structs":
| commit b11a04266088df7dec1bf9e8ee2db96cd641578b
| Author: Zachary Leaf <zachary.leaf@arm.com>
| Date:   Wed Nov 2 10:13:25 2022 +0000
|
|     bpf: use user pointer types in uAPI structs
|
|     union bpf_attr and structs bpf_xyz_info are used to pass config to the
|     kernel or return info to userspace via the bpf syscall.
|
|     In order to avoid needing a separate 32-bit compat handler for the bpf
|     syscall, __user pointers inside these union/structs are stored as __u64
|     or __aligned_u64.
|
|     In PCuABI, a user pointer is a 129-bit capability so __aligned_u64 type
|     is not big enough to hold it. Use type __kernel_aligned_uintptr_t
|     instead, which is big enough on the affected architectures while
|     remaining __aligned_64 on others.
|
|     Use copy_{to,from}_user_with_ptr variants where blocks of memory
|     containing pointers are being copied to preserve capabilities.
|
|     Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>

+++ kernel/bpf/verifier.c
@@ -25998,8 +25998,7 @@ int bpf_check(struct bpf_prog **prog, union bpf_attr *attr, bpfptr_t uattr, __u3
 	 * and supplied buffer to store the verification trace
 	 */
 	ret = bpf_vlog_init(&env->log, attr->log_level,
-			    (char __user *) (unsigned long) attr->log_buf,
-			    attr->log_size);
+			    (char __user *)attr->log_buf, attr->log_size);
 	if (ret)
 		goto err_unlock;
